server-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain>=0.0.300
langchain-openai>=0.0.2
openai>=1.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
crawl4ai==0.4.248
